config adjust to yaml
