+++ cmd/callisto/config.yaml
@@ -1,10 +1,10 @@
 server:
-  name: "amusing-x/pluto"
+  name: "amusing-x/callisto"
   httpServer:
-    addr: :10003
+    addr: :10002
   grpcServer:
     network: tcp
-    address: :20003
+    address: :20002
 mysql:
   plutodb:
     db: plutodb
@@ -20,3 +20,8 @@ redis:
   redis1:
     addr: 127.0.0.1:6379
     dbNo: 1
+session:
+  redisStrore:
+    redisAddr: localhost:6379
+    redisNo: 2
+    keyPrefix: uid
