--- cmd/callisto/service_testing.conf
@@ -1,27 +1,0 @@
-[base]
-addr :10002
-http.port :10002
-
-mysql.callisto.database amusingrisk
-mysql.callisto.username amusing
-mysql.callisto.password amusing111
-mysql.callisto.host localhost
-mysql.callisto.port 3306
-mysql.callisto.protocol tcp
-mysql.callisto.connMaxLifetime 3600
-mysql.callisto.maxOpenConns 100
-mysql.callisto.maxIdleConns 10
-
-redis.addr localhost:6379
-redis.db 0
-
-[rpc]
-network tcp
-address localhost:11002
-
-[session]
-store redis
-store.redis.addr localhost:6379
-# store.redis.password
-store.redis.db 2
-store.redis.key.prefix uid
