--- cmd/charon/service_testing.conf
@@ -1,19 +1,0 @@
-[base]
-addr :10004
-http.port :10004
-server.name :charonService
-
-[redis]
-redis.addr 192.168.56.1:6379
-redis.db 0
-
-[rpc]
-network tcp
-address :11004
-
-[session]
-store redis
-store.redis.addr localhost:6379
-# store.redis.password
-store.redis.db 2
-store.redis.key.prefix uid
