--- cmd/europa/service_testing.conf
@@ -1,28 +1,0 @@
-[base]
-addr :10003
-http.port :10003
-
-mysql.europa.database amusinguser
-mysql.europa.username amusing
-mysql.europa.password amusing111
-mysql.europa.host localhost
-mysql.europa.port 3306
-mysql.europa.protocol tcp
-mysql.europa.connMaxLifetime 3600
-mysql.europa.maxOpenConns 100
-mysql.europa.maxIdleConns 10
-
-redis.addr localhost:6379
-redis.db 0
-
-[rpc]
-amusingriskservice.address localhost:11002
-amusingxuserservice.address localhost:11001
-network tcp
-
-[session]
-store redis
-store.redis.addr localhost:6379
-# store.redis.password
-store.redis.db 2
-store.redis.key.prefix uid
