+++ cmd/ganymede/config.yaml
@@ -1,10 +1,10 @@
 server:
   name: "amusing-x/pluto"
   httpServer:
-    addr: :10003
+    addr: :10005
   grpcServer:
     network: tcp
-    address: :20003
+    address: :20005
 mysql:
   plutodb:
     db: plutodb
