--- cmd/ganymede/service_testing.conf
@@ -1,28 +1,0 @@
-[base]
-addr :10001
-http.port :10001
-
-mysql.ganymede.database amusinguser
-mysql.ganymede.username amusing
-mysql.ganymede.password amusing111
-mysql.ganymede.host localhost
-mysql.ganymede.port 3306
-mysql.ganymede.protocol tcp
-mysql.ganymede.connMaxLifetime 3600
-mysql.ganymede.maxOpenConns 100
-mysql.ganymede.maxIdleConns 10
-
-redis.addr localhost:6379
-redis.db 0
-
-[rpc]
-amusingriskservice.address localhost:11002
-amusingxuserservice.address localhost:11001
-network tcp
-
-[session]
-store redis
-store.redis.addr localhost:6379
-# store.redis.password
-store.redis.db 2
-store.redis.key.prefix uid
