--- cmd/pluto/service_testing.conf
@@ -1,30 +1,0 @@
-[base]
-addr :10003
-http.port :10003
-server.name :plutoService
-
-[mysql]
-mysql.pluto.db plutodb
-mysql.pluto.username test111
-mysql.pluto.password pass123456
-mysql.pluto.host localhost
-mysql.pluto.port 3306
-mysql.pluto.protocol tcp
-mysql.pluto.connMaxLifetime 3600
-mysql.pluto.maxOpenConns 100
-mysql.pluto.maxIdleConns 10
-
-[redis]
-redis.addr 127.0.0.1:6379
-redis.db 0
-
-[rpc]
-network tcp
-address :11002
-
-[session]
-store redis
-store.redis.addr localhost:6379
-# store.redis.password
-store.redis.db 2
-store.redis.key.prefix uid
