+++ services/charon/conf/conf.go
@@ -2,48 +2,70 @@ package conf
 
 import (
 	"github.com/ItsWewin/superfactory/basicmatter"
-	"log"
-	"reflect"
+	"github.com/ItsWewin/superfactory/logger"
 )
 
-var Conf2 = new(ConfigSection)
+var Conf = new(Config)
 
-type ConfigSection struct {
-	Addr       string `config:"base:addr"`
-	Port       string `config:"base:http.port"`
-	ServerName string `config:"base:server.name"`
-
-	RedisAddr     string `config:"redis:redis.addr"`
-	RedisPassword string `config:"redis:redis.password"`
-	RedisDB       int    `config:"redis:redis.db"`
-
-	SessionStore               string `config:"session:store"`
-	SessionStoreRedisAddr      string `config:"session:store.redis.addr"`
-	SessionStoreRedisPassword  string `config:"session:store.redis.password"`
-	SessionStoreRedisDB        int    `config:"session:store.redis.db"`
-	SessionStoreRedisKeyPrefix string `config:"session:store.redis.key.prefix"`
-
-	RPCNetwork string `config:"rpc:network"`
-	RPCAddress string `config:"rpc:address"`
+type Config struct {
+	Server *Server   `yaml:"server"`
+	Mysql  MysqlConf `yaml:"mysql"`
+	Redis  Redis     `yaml:"redis"`
 }
 
-func (c *ConfigSection) MaterType() string {
-	return basicmatter.MasterConfigBasicSection
+type Server struct {
+	Name       string      `yaml:"name"`
+	HttpServer *HttpServer `yaml:"httpServer"`
+	GrpcServer *GrpcServer `yaml:"grpcServer"`
 }
 
-func (c *ConfigSection) HttpAddr() string {
-	return c.Addr
+type HttpServer struct {
+	Addr string `yaml:"addr"`
 }
 
-func (c *ConfigSection) Print() {
-	rv := reflect.ValueOf(*c)
-	rt := reflect.TypeOf(*c)
+type GrpcServer struct {
+	Network string `yaml:"network"`
+	Address string `yaml:"address"`
+}
 
-	log.Println("==================== service config ====================")
-	for i := 0; i < rv.NumField(); i++ {
-		if rt.Field(i).Name == "Port" {
-			log.Printf("%v:%v\n", rt.Field(i).Name, rv.Field(i))
-		}
-	}
-	log.Println("==================== service config ====================")
+type MysqlConf struct {
+	Plutodb *Mysql `yaml:"plutodb"`
+}
+
+type Mysql struct {
+	DB           string `yaml:"db"`
+	User         string `yaml:"user"`
+	Password     string `yaml:"password"`
+	Host         string `yaml:"host"`
+	Port         string `json:"port"`
+	Protocol     string `yaml:"protocol"`
+	MaxLifeTime  int64  `yaml:"maxLifeTime"`
+	MaxOpenConns int    `yaml:"maxOpenConns"`
+	MaxIdleConns int    `yaml:"maxIdleConns"`
+}
+
+type Redis struct {
+	RedisO RedisConf `yaml:"redis_0"`
+}
+
+type RedisConf struct {
+	Addr     string `yaml:"addr"`
+	DBNo     int    `yaml:"dbNo"`
+	Password string `yaml:"password"`
+}
+
+func (c *Config) MaterType() string {
+	return basicmatter.MasterConfigBasicYaml
+}
+
+func (c *Config) RunHttpServer() bool {
+	return true
+}
+
+func (c *Config) HttpAddr() string {
+	return c.Server.HttpServer.Addr
+}
+
+func (c *Config) Print() {
+	logger.Infof("config: %s", logger.ToJson(c))
 }
